refactor(go-mod): Moved GoModWhyParser to parse package.
