+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/go/gomod/parse/GoModWhyParser.java
@@ -5,7 +5,7 @@
  *
  * Use subject to the terms and conditions of the Synopsys End User Software License and Maintenance Agreement. All rights reserved worldwide.
  */
-package com.synopsys.integration.detectable.detectables.go.gomod;
+package com.synopsys.integration.detectable.detectables.go.gomod.parse;
 
 import java.util.LinkedHashSet;
 import java.util.List;
